arranged classes into more organized packages
