+++ src/Buzzword/Model/CreateProfileException.java
@@ -1,4 +1,4 @@
-package Buzzword;
+package Buzzword.Model;
 
 /**
  * Created by Phil on 11/26/2016.
