+++ src/Buzzword/View/AppMessageDialogSingleton.java
@@ -1,4 +1,4 @@
-package Buzzword;
+package Buzzword.View;
 
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
